azure_File-portal/requirements.txt
flask
azure-storage-blob
azure-identity
gunicorn
